requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Gradio UI
gradio==4.16.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Database
qdrant-client==1.7.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# Email (AWS SES)
boto3==1.34.34

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# PII/PHI Protection
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
spacy==3.7.2

# AI/LLM Providers
openai==1.10.0
anthropic==0.9.0
google-generativeai==0.3.2
langchain==0.1.4
langchain-openai==0.0.5

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3

# Logging & Monitoring
loguru==0.7.2